third_party/WebKit/LayoutTests/external/wpt/content-security-policy/navigate-to/link-click-blocked-expected.txt

This is a testharness.js-based test.
FAIL Test that the child iframe navigation is blocked assert_equals: expected "fail" but got "success"
Harness: the test ran to completion.
